# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "族谱管理系统的RESTful API文档",
        "title": "族谱管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/person/create": {
            "post": {
                "description": "创建新的家族成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "创建人员",
                "parameters": [
                    {
                        "description": "人员信息",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/delete": {
            "delete": {
                "description": "删除家族成员",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "删除人员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人员ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/getById": {
            "get": {
                "description": "根据ID获取家族成员信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "根据ID获取人员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人员ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/getByUuid": {
            "get": {
                "description": "根据UUID获取家族成员信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "根据UUID获取人员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "人员UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/list": {
            "get": {
                "description": "分页列出家族成员",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "列出人员",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/search": {
            "get": {
                "description": "根据姓名搜索家族成员",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "搜索人员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/person/update": {
            "put": {
                "description": "更新家族成员的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人员管理"
                ],
                "summary": "更新人员信息",
                "parameters": [
                    {
                        "description": "人员信息",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "person.Person": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "biography": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_lunar": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "birth_place_latitude": {
                    "type": "number"
                },
                "birth_place_longitude": {
                    "type": "number"
                },
                "cover_photo_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "death_cause": {
                    "type": "string"
                },
                "death_date": {
                    "type": "string"
                },
                "death_lunar": {
                    "type": "string"
                },
                "death_place": {
                    "type": "string"
                },
                "death_place_latitude": {
                    "type": "number"
                },
                "death_place_longitude": {
                    "type": "number"
                },
                "education": {
                    "type": "string"
                },
                "english_name": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "generation_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_alive": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "verification_status": {
                    "type": "string"
                }
            }
        },
...
    }
}